Atributo mundos añadidos al User y selector de mundos modificado

+++ app/src/main/java/edu/ub/pis/joc/limitless/model/User.kt
@@ -7,12 +7,12 @@ class User : Serializable {
     var realName: String? = null
     var email: String? = null
     var survived: Long? = null
-    var world: Long? = null
+    var world: Int? = null
 
     constructor() {
     }
 
-    constructor(userName: String, realName: String, email: String, survived: Long, world: Long) {
+    constructor(userName: String, realName: String, email: String, survived: Long, world: Int) {
         this.userName = userName
         this.realName = realName
         this.email = email
